pyproject.toml
[project]
name = "task-manager"
version = "0.1.0"
description = "Celery-based distributed task framework with GPU worker support"
requires-python = "==3.10.*"
dependencies = [
    "celery[redis]>=5.3.0",
    "redis>=5.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "docker>=7.0.0",
    "prometheus-client>=0.19.0",
    "httpx>=0.25.0",
    "pillow>=10.1.0",
    "aiofiles>=23.2.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"